Fix logging of couchdb errors.
Leap MX was previously wrapping soledad's couchdb-related exceptions
messages in its own exception. The problem was that message attribute of
those exceptions were actually empty, and so the information of the
actual error was lost. This commit removes the wrapper exception and
forwards whatever exception was received from soledad to the logging
system.

Signed-off-by: Ruben Pollan <meskio@sindominio.net>