[bug] check if key is expired
For expired keys gnupg will not allow us to encrypt (to force it we'll
need to fake the system date). Let's bounce the email properly in this case,
and register the error.

Make sure as well that _encrypt_message does not return None, let's
raise exceptions in case of failure.

- Related: #8663